+++ DTA_Module.py
@@ -16,7 +16,7 @@ class TemporalAttention(nn.Module):
   '''
   def __init__(self, device, d):
     super().__init__()
-    self.w_ta =  nn.Parameter(torch.randn((d, d), requires_grad=True, device=device, dtype=torch.bfloat16)) * 0.01
+    self.w_ta =  nn.Parameter(torch.randn((d, d), requires_grad=True, device=device, dtype=torch.float32)) * 0.01
   def forward(self, h_t_prev):
     beta_t = nn.Tanh()(h_t_prev)
     beta_t = torch.matmul(self.w_ta, h_t_prev)
